Add progress on converting between new note datatypes.
This commit leaves the repository in a non-working state. Additional
functionality for `_pull()` must be implemented in order to make use of
the SQL parser.
